app/requirements.txt
flask
mysql-connector
folium